RAG_Pilot/requirements.txt
sentence-transformers==2.7.0
faiss-cpu>=1.7.4
python-docx>=1.1.0
pypdf>=4.2.0
numpy>=1.24
rich>=13.7.0
